added header styles and responsive logout button
