+++ react-app/public/index.html
@@ -2,7 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
+    <meta content="width=device-width, initial-scale=1" name="viewport" />
     <title>Breedapedia</title>
+    <script src="https://kit.fontawesome.com/de73546d8a.js" crossorigin="anonymous"></script>
   </head>
   <body>
     <div id="root"></div>
