+++ react-app/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import LoginForm from './components/auth/LoginForm';
 import SignUpForm from './components/auth/SignUpForm';
-import NavBar from './components/NavBar';
+import Header from './components/header/Header';
 import ProtectedRoute from './components/auth/ProtectedRoute';
 import UsersList from './components/UsersList';
 import User from './components/User';
@@ -27,7 +27,7 @@ function App() {
 
   return (
     <BrowserRouter>
-      <NavBar />
+      <Header />
       <Switch>
         <Route path='/login' exact={true}>
           <LoginForm />
