+++ react-app/src/components/auth/LogoutButton.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { logout } from '../../store/session';
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faRightFromBracket } from "@fortawesome/free-solid-svg-icons";
+import './auth.css'
 
 const LogoutButton = () => {
   const dispatch = useDispatch()
@@ -8,7 +11,11 @@ const LogoutButton = () => {
     await dispatch(logout());
   };
 
-  return <button onClick={onLogout}>Logout</button>;
+  return (
+    <button className='logout__button' onClick={onLogout}>
+      <FontAwesomeIcon icon={faRightFromBracket} />
+    </button>
+  )
 };
 
 export default LogoutButton;
