+++ react-app/src/components/auth/auth.css
@@ -1,0 +1,18 @@
+/* Logout Styles */
+
+.logout__button {
+    background-color: transparent;
+    border-width: 0;
+    font-family: inherit;
+    font-size: inherit;
+    font-style: inherit;
+    font-weight: inherit;
+    line-height: inherit;
+    padding: 0;
+    color: white;
+}
+
+.logout__button:hover {
+    color: var(--gold);
+    cursor: pointer;
+}
